requirements.txt
azure-functions
PyPDF2
pdf2image
pytesseract
Pillow
